feat(plugin): add batch write with action

+++ plugin/plugin.go
@@ -25,7 +25,7 @@ type Adapter interface {
 type BatchAdapter interface {
 	Adapter
 	PutBlocks(ctx context.Context, blks []*block.Block) error
-	BatchSize() uint64
+	BatchSize() int
 }
 
 type DependentAdapter interface {
