+++ plugins/account_income_v1/account_income.go
@@ -45,8 +45,8 @@ func (b accountIncomeV1Plugin) Type() plugin.Type {
 	return plugin.TypeStandard
 }
 
-func (b accountIncomeV1Plugin) BatchSize() uint64 {
-	return uint64(b.batchSize)
+func (b accountIncomeV1Plugin) BatchSize() int {
+	return b.batchSize
 }
 
 func (b *accountIncomeV1Plugin) Start(ctx context.Context) error {
